requirements.txt
django
djangorestframework
celery
redis
langchain
langchain-groq
python-dotenv
gunicorn
whitenoise